# Conflicts:
#   app/app/src/main/java/com/bwabwayo/app/domain/chat/service/ChatRoomService.java

+++ app/app/src/main/java/com/bwabwayo/app/global/storage/util/StorageUtil.java
@@ -13,9 +13,13 @@ import java.util.List;
 @Component
 @RequiredArgsConstructor
 public class StorageUtil {
+    private final boolean ENABLE_REALTIME_DELETE = false;
+
     private final StorageService storageService;
 
     public void deleteWithoutException(String key) {
+        if(!ENABLE_REALTIME_DELETE) return;
+
         try {
             storageService.delete(key);
         } catch (Exception e) {
@@ -40,7 +44,7 @@ public class StorageUtil {
     public String copyToDirectory(String key, String srcPrefix, String tgtPrefix) {
         if (!storageService.exists(key)) {
             log.info("스토리지에 존재하지 않는 파일: key={}", key);
-            throw new NotFoundException("스토리지에 존재하지 않는 파일: key=" + key);
+            return key;
         }
 
         String tgtKey = key;
@@ -55,12 +59,16 @@ public class StorageUtil {
      * keys 중 prefix가 srcPrefix가 tgtPrefix로 복사된 key를 삭제
      */
     public void rollback(List<String> keys, String srcPrefix, String tgtPrefix){
+        if(!ENABLE_REALTIME_DELETE) return;
+
         for(String key : keys){
             rollback(key, srcPrefix, tgtPrefix);
         }
     }
 
     public void rollback(String key, String srcPrefix, String tgtPrefix){
+        if(!ENABLE_REALTIME_DELETE) return;
+
         if(key.startsWith(srcPrefix)){
             String deletingKey = tgtPrefix + key.substring(srcPrefix.length());
             deleteWithoutException(deletingKey);
